+++ src/cieinr/cieinr_test_linkml.json
@@ -1,0 +1,112 @@
+[
+  {
+    "record_id": "16",
+    "basic_form": {
+      "iei_deficiency_basic": "mondo_0007843",
+      "basic_form_complete": "2",
+      "igrt_basic": "ncit_c62710",
+      "hct_basic_form": "ncit_c15431"
+    },
+    "patient_demographics_initial_form": {
+      "visit_date_demographics": "2025-03-15",
+      "snomedct_184099003": "2022-03-01",
+      "snomedct_432213005": "",
+      "snomedct_298059007": "",
+      "patient_demographics_initial_form_complete": "2"
+    },
+    "repeated_elements": [
+      {
+        "redcap_repeat_instrument": "infections_initial_form",
+        "redcap_repeat_instance": 1,
+        "infections_initial_form": {
+          "type_of_infection": "snomedct_21483005",
+          "snomedct_61274003": "",
+          "snomedct_21483005": "hp_0002383",
+          "snomedct_81745001": "",
+          "snomedct_385383008": "",
+          "snomedct_127856007": "",
+          "snomedct_110522009": "",
+          "snomedct_20139000": "",
+          "snomedct_303699009": "",
+          "snomedct_21514008": "",
+          "snomedct_31099001": "",
+          "infection_severity": "hp_0012826",
+          "infection_temp_pattern": "hp_0011009",
+          "infection_times_obseverd": "ncit_c156502",
+          "infections_initial_form_complete": "2"
+        }
+      },
+      {
+        "redcap_repeat_instrument": "infections_initial_form",
+        "redcap_repeat_instance": 2,
+        "infections_initial_form": {
+          "type_of_infection": "snomedct_127856007",
+          "snomedct_61274003": "",
+          "snomedct_21483005": "",
+          "snomedct_81745001": "",
+          "snomedct_385383008": "",
+          "snomedct_127856007": "mondo_0043653",
+          "snomedct_110522009": "",
+          "snomedct_20139000": "",
+          "snomedct_303699009": "",
+          "snomedct_21514008": "",
+          "snomedct_31099001": "",
+          "infection_severity": "hp_0012826",
+          "infection_temp_pattern": "hp_0011011",
+          "infection_times_obseverd": "ncit_c156502",
+          "infections_initial_form_complete": "0"
+        }
+      },
+      {
+        "redcap_repeat_instrument": "infections_initial_form",
+        "redcap_repeat_instance": 3,
+        "infections_initial_form": {
+          "type_of_infection": "snomedct_20139000",
+          "snomedct_61274003": "",
+          "snomedct_21483005": "",
+          "snomedct_81745001": "",
+          "snomedct_385383008": "",
+          "snomedct_127856007": "",
+          "snomedct_110522009": "",
+          "snomedct_20139000": "hp_0002090",
+          "snomedct_303699009": "",
+          "snomedct_21514008": "",
+          "snomedct_31099001": "",
+          "infection_severity": "hp_0012826",
+          "infection_temp_pattern": "hp_0011009",
+          "infection_times_obseverd": "ncit_c65134",
+          "infections_initial_form_complete": "2"
+        }
+      },
+      {
+        "redcap_repeat_instrument": "genetic_information",
+        "redcap_repeat_instance": 1,
+        "genetic_findings": {
+          "g_iei_deficiency": "mondo_0011383",
+          "other_non_g_iei": "",
+          "g_other_iei_deficiency": "",
+          "ga4gh_progress_status": "",
+          "ga4gh_interp_status": "",
+          "loinc_81304_8": "loinc_la46-8",
+          "loinc_62374_4": "",
+          "loinc_lp7824_8": "",
+          "variant_expression": "chgvs",
+          "loinc_81290_9": "",
+          "loinc_48004_6": "NM_032977.3:c.-360T>C",
+          "loinc_48005_3": "",
+          "variant_validation": "yes",
+          "loinc_48018_6": "HGNC:1500",
+          "loinc_53034_5": "loinc_la6706-1",
+          "loinc_53034_5_other": "",
+          "loinc_48002_0": "",
+          "loinc_48019_4": "loinc_la6688-1",
+          "loinc_48019_4_other": "",
+          "loinc_53037_8": "loinc_la26333-7",
+          "ga4gh_therap_action": "",
+          "loinc_93044_6": "",
+          "rarelink_6_1_genetic_findings_complete": "0"
+        }
+      }
+    ]
+  }
+]
